41 added pipeline options

+++ mproc-pipeline/batch.sh
@@ -9,29 +9,52 @@
 # (bloc.h is set to ROWS_R_SNPS=0 in mproc project folder)
 # (BlocBuster executable files are compiled to return 0 as true rather than 1 for Linux responses)
 
-case_file="../data/TSI-8-102-1000" # path to file outside data folder
-control_file="../data/TSI-8-102-1000" # path to file outside data folder
-blocbuster_path="/home/jjs3k2/mproc/mproc/blocbuster" # path to executable file within mproc folder
+# please specify paths, either relative or absolute
+# these should already work correctly
+case_file="../data/example_102_1000.txt" # path to file outside data folder
+control_file="../data/example_102_1000.txt" # path to file outside data folder
+blocbuster_path="../mproc/blocbuster" # path to executable file within mproc folder
+keephi_path="../blocbuster/keepHi/keepHi" # path to executable file within blocbuster folder
+bfs_path="../blocbuster/bfs/bfs" # path to executable file within blocbuster folder
+carriers_path="../blocbuster/carriers/carriers" # path to executable file within blocbuster folder
+# please specify file names, not paths
 blocbuster_folder="log" # folder name without path, builds a temp folder within the data folder
-bfs_path="/home/jjs3k2/BlocBuster/bfs/bfs" # path to executable file within BlocBuster folder
-carriers_path="/home/jjs3k2/BlocBuster/carriers/carriers" # path to executable file within BlocBuster folder
 gml_file="out.gml" # file name without path
+keephi_file="keephi.gml" # file name without path
 bfs_file="out.bfs" # file name without path
 carriers_file="carriers.out" # file name without path
 snp_info_file="snps.info" # file name without path
+# please specify strings, integers, or floating point values
 delimiter=' ' # delimiter in case/control file
 let num_cases=102
 let num_controls=102
 let num_snps=1000
 let case_control_header_rows=1
 let case_control_header_columns=11
-threshold="0.7"
-let granularity1=0
+threshold=0.7
+edges_to_keep=0.5
+let granularity1=3
 let granularity2=3
-let maxProcs=6
+let maxProcs=10
 let semaphores=0 # 0 = false, 1 = true, default 0 (recommended)
 
 # no modification required after this line
+
+skip="none"
+if [[ $# -ge 1 ]]
+then
+	if [[ $1 == "-h" ]]
+	then
+		echo "usage: batch.sh [-h -a -z]"
+		exit 0
+	elif [[ $1 == "-a" ]]
+	then
+		skip="last"
+	elif [[ $1 == "-z" ]]
+	then
+		skip="first"
+	fi
+fi
 
 # data cleaning
 let info_header_columns=$case_control_header_columns
@@ -41,10 +64,18 @@ then
 	echo "error - fewer snps than individuals"
 	exit 1
 fi
+if [[ ! -e $case_file || ! -e $control_file || ! -e $blocbuster_path || ! -e $keephi_path || ! -e $bfs_path || ! -e $carriers_path ]]
+then
+	echo "error - could not find one of the input files"
+	exit 1
+fi
 tmp_dir="data"
-rm -rf $tmp_dir
-sleep 1
-mkdir $tmp_dir
+if [[ $skip != "first" ]]
+then
+	rm -rf $tmp_dir
+	sleep 1
+	mkdir $tmp_dir
+fi
 tmp_file="tmp.txt"
 tmp_file2="tmp2.txt"
 tmp_file3="tmp3.txt"
@@ -57,6 +88,10 @@ snp_info_file2="$tmp_dir/$snp_info_file"
 snp_info_file=$snp_info_file2
 
 function clean(){
+	if [[ $skip == "first" ]]
+	then
+		return
+	fi
 	infile=$1
 	echo "processing $infile"
 	resultfile=$2
@@ -127,7 +162,7 @@ infile=$case_file
 outfile=$gml_file
 let num_ind=$num_cases
 
-if [[ $granularity1 -gt 0 ]]
+if [[ $skip != "first" && $granularity1 -gt 0 ]]
 then
 	pidlist=`srun mproc1.sh $blocbuster_path $infile $outfile $threshold $num_ind $num_snps $case_control_header_rows $case_control_header_columns $granularity1 $granularity2 $maxProcs $blocbuster_folder $semaphores`
 	out=${pidlist#*Submitted batch job}
@@ -141,13 +176,14 @@ then
 	done
 	pidlist=`echo $ids | tr ' ' ','`
 	echo "running sbatch $pidlist"
-else
+elif [[ $skip != "first" ]]
+then
 	echo "************************************"
 	echo "running mproc"
 	bash virt_mproc1.sh $blocbuster_path $infile $outfile $threshold $num_ind $num_snps $case_control_header_rows $case_control_header_columns $granularity1 $granularity2 $maxProcs $blocbuster_folder $semaphores
 fi
 
-if [[ $granularity1 -gt 0 ]]
+if [[ $skip != "first" && $granularity1 -gt 0 ]]
 then
 	echo "waiting for processes to complete"
 	for p in ${pids[@]}
@@ -171,15 +207,15 @@ then
 			echo "process $p failed"
 			exit 1
 		fi
-		echo "process $p complete"
-done
+		echo "process $p completed"
+	done
 fi
 
 # mproc 
 
 #params="--dependency=afterok:$pidlist mproc2.sh"
 #echo $params
-if [[ $granularity1 -gt 0 ]]
+if [[ $skip != "first" && $granularity1 -gt 0 ]]
 then
 	params="mproc2.sh $blocbuster_path"
 	x=`srun $params`
@@ -187,16 +223,66 @@ then
 	pid=`echo $y | sed "s/ //g"`
 	echo "running sbatch $pid"
 	pid1=$pid
-else
+elif [[ $skip != "first" ]]
+then
 	params="virt_mproc2.sh $blocbuster_path"
 	echo "************************************"
 	echo "running mproc -z"
 	bash $params
 fi
 
+if [[ $skip == "last" ]]
+then
+	echo "waiting for processes to complete"
+	params="sacct -j $pid --format=state"
+	stats=`$params`
+	let stat=`echo $stats | grep "COMPLETED" | wc -l`
+	let stat2=`echo $stats | grep "FAILED" | wc -l`
+	while [[ $stat -eq 0 && $stat2 -eq 0 ]]
+	do
+		sleep 1
+		stats=`$params`
+		let stat=`echo $stats | grep "COMPLETED" | wc -l`
+		let stat2=`echo $stats | grep "FAILED" | wc -l`
+		#echo $stat
+	done
+	if [[ $stat2 -ne 0 ]]
+	then
+		echo "process $p failed"
+		exit 1
+	fi
+	echo "process $p completed"
+	exit 0
+fi
+
+# keepHi
+
+infile="$tmp_dir/$gml_file"
+outfile="$tmp_dir/$keephi_file"
+let nodes=$(( num_snps * 2 ))
+let edges=0
+
+if [[ $granularity1 -gt 0 ]]
+then
+	options=""
+	if [[ $skip != "first" ]]
+	then
+		options="--dependency=afterok:$pid"
+	fi
+	params="--parsable $options keephi.sh $keephi_path $infile $nodes $edges $edges_to_keep $outfile"
+	pid=`sbatch $params`
+	echo "running sbatch $pid"
+	pid2=$pid
+else
+	echo "************************************"
+	echo "running keepHi"
+	params="virt_keephi.sh $keephi_path $infile $nodes $edges $edges_to_keep $outfile"
+	bash $params
+fi
+
 # bfs
 
-infile="$tmp_dir/$outfile"
+infile="$tmp_dir/$keephi_file"
 outfile="$tmp_dir/$bfs_file"
 
 if [[ $granularity1 -gt 0 ]]
@@ -204,7 +290,7 @@ then
 	params="--parsable --dependency=afterok:$pid bfs.sh $bfs_path $infile $outfile"
 	pid=`sbatch $params`
 	echo "running sbatch $pid"
-	pid2=$pid
+	pid3=$pid
 else
 	echo "************************************"
 	echo "running bfs"
@@ -221,7 +307,7 @@ if [[ $granularity1 -gt 0 ]]
 then
 	pid=`sbatch --parsable --dependency=afterok:$pid carriers.sh $carriers_path $infile $case_file $control_file $case_control_header_rows $case_control_header_columns $snp_info_file $info_header_columns $info_header_rows $num_cases $num_controls $num_snps $outfile`
 	echo "running sbatch $pid"
-	pid3=$pid
+	pid4=$pid
 else
 	echo "************************************"
 	echo "running carriers"
@@ -241,7 +327,7 @@ then
 		let stat=`echo $stats | grep "COMPLETED" | wc -l`
 		let stat2=`echo $stats | grep "COMPLETED" | wc -l`
 	done
-	echo "$pid1,$pid2,$pid3 complete"
+	echo "$pid1,$pid2,$pid3,$pid4 completed"
 else
 	echo "************************************"
 	echo "complete"
